+++ test/HelpTest.php
@@ -1,0 +1,61 @@
+<?php
+/**
+ * @license   http://opensource.org/licenses/BSD-3-Clause BSD-3-Clause
+ * @copyright Copyright (c) 2017 Zend Technologies USA Inc. (http://www.zend.com)
+ */
+
+namespace ZFTest\ComposerAutoloading;
+
+use PHPUnit\Framework\TestCase;
+use Prophecy\Argument;
+use Zend\Stdlib\ConsoleHelper;
+use ZF\ComposerAutoloading\Help;
+
+class HelpTest extends TestCase
+{
+    public function testWritesHelpMessageToConsoleUsingCommandProvidedAtInstantiationAndResourceAtInvocation()
+    {
+        $resource = fopen('php://temp', 'wb+');
+
+        $console = $this->prophesize(ConsoleHelper::class);
+        $console
+            ->writeLine(
+                Argument::that(function ($message) {
+                    return false !== strpos($message, 'zf-composer-autoloading');
+                }),
+                true,
+                $resource
+            )
+            ->shouldBeCalled();
+
+        $command = new Help(
+            'zf-composer-autoloading',
+            $console->reveal()
+        );
+
+        $this->assertNull($command($resource));
+    }
+
+    public function testTruncatesCommandToBasenameIfItIsARealpath()
+    {
+        $resource = fopen('php://temp', 'wb+');
+
+        $console = $this->prophesize(ConsoleHelper::class);
+        $console
+            ->writeLine(
+                Argument::that(function ($message) {
+                    return false !== strpos($message, basename(__FILE__));
+                }),
+                true,
+                $resource
+            )
+            ->shouldBeCalled();
+
+        $command = new Help(
+            realpath(__FILE__),
+            $console->reveal()
+        );
+
+        $this->assertNull($command($resource));
+    }
+}